// src/app/models/userProv.ts
export class UserProv {
  name: string;
  surname: string;
  surname2: string;
  email: string;
  password: string;
  eid: string;
  adress: string;
  city: string;
  postalcode: string;
  country: string;
  phone: string;
  phone2: string;
  mobilePhone:string;
  cif: string;
  referenceCode :string;
  internalNotes:string;
  lote:string;
  suites: string[];
  partnerInternalReference :string;
  parentEid:string;
  parentEmail:string;
  parentCompanyIdentification:string;
  companyName: string;
  companyIdentification: string;
  companyAddress: string;
  companyPhones: string;
  companyAuthorizedFullname: string;
  companyAuthorizedPhones: string;
  companyMaxNumLicensesAllowed:number;
  externalLicenses:string;
  expirationDate: Date;
  activationDate: Date;

}