// src/app/models/userChildProv.ts
export class UserChildProv {
  name: string;
  surname: string;
  email: string;
  eid: string;
  address: string;
  city: string;
  postalcode: string;
  countryCode: string;
  country: string;
  phone: string;
  phone2: string;
  profileImage:string;
  suites: string[];
  parentEid:string;
  parentEmail:string;
  parentCompanyIdentification:string;
  adminUserId:number;
  mobilePhone:string;
  cif: string;
}